Call State rearch (#598)
* From fork

* Bump version

* Review
