+++ Decimus/Codec/LibOpusDecoder.swift
@@ -5,16 +5,18 @@ import Opus
 import AVFAudio
 
 /// Decodes audio using libopus.
-class LibOpusDecoder {
+class LibOpusDecoder: AudioDecoder {
     private static let logger = DecimusLogger(LibOpusDecoder.self)
 
     private let decoder: Opus.Decoder
     let decodedFormat: AVAudioFormat
+    let encodedFormat: AVAudioFormat
 
     /// Create an opus decoder.
     /// - Parameter format: Format to decode into.
     init(format: AVAudioFormat) throws {
         self.decodedFormat = format
+        self.encodedFormat = format
         decoder = try .init(format: format, application: .voip)
     }
