+++ Decimus/Publications/AudioPublication.swift
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025 Cisco Systems
+// SPDX-License-Identifier: BSD-2-Clause
+
+protocol AudioPublication {
+    func togglePublishing(active: Bool)
+}
